+++ CHANGELOG.md
@@ -43,3 +43,4 @@
    - FIX: re-implement changing of list keys by using yang schema on \__setattr__ rather than trusting the backend to enforce it.
    - FIX: sysrepo dataa-abstraction layer dirty now captures if we have made changes (raises NoChangesToCommit) if there are no changes.
    - FEATURE: stubdal now provides default values
+   - REMOVE: un-used module from list create operation